list.rs and stmts.rs split under the 800-line cap: flat_map / filter_map and the index-assign route move to their own modules
