+++ crates/almide-wasm/src/lib.rs
@@ -126,6 +126,7 @@ pub(crate) use func::*;
 pub(crate) mod ty;
 pub(crate) use ty::*;
 mod list;
+mod list_flat;
 mod list_comb;
 mod list_edit;
 mod list_search;
@@ -135,6 +136,7 @@ mod list_order;
 mod list_sort;
 mod string_scan;
 mod stmts;
+mod stmts_index;
 mod stmts_append;
 mod string_ext;
 pub(crate) mod work;
